# Api/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/pos/v0.1/Reports/SalesByRep": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReportSalesByRepRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/SalesDetailsByPaymentMode": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SalesDetailsByPaymentModeRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/TransactionDetails": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TrnDetReportRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/SalesRepWiseCustomer": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SalesRepWiseCustomerRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/SalesItemReport": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SalesItemReportRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/NonPerformedItems": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/NonPerformedItemRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/NonPerformedCustomers": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/NonPerformedCustomerRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/ChequeReturnDetails": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ChqReturnReportRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/AccountLedger": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AccLedgerReportRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/AccountTransactionDetails": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AccTrnDetailsReportRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/AccountBalanceAsAt": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AccBalanceAsAtRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/ItemCat1WiseTransactionSummary": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ItmCat1WiseTrnSumRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/SetOffDetails": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SetOffDetailsReportRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/GrossProfit": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/GrossProfitItemLocationRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/CreditSalesSummary": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LssCreditInvoiceSummaryRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/AgeAnalysis": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AgeAnalysisRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/DueStatement": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DebtorsDueStatementRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/StockLedger": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StockLedgerRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/StockAsAt": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StockAsAtRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/ItemBatchDetails": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ItemBatchReportRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/Reports/ReOrderItems": {
      "get": {
        "tags": [
          "Reports"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReOrderItemsRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AccBalanceAsAtRequestDto": {
      "type": "object",
      "properties": {
        "asAtDate": {
          "format": "date",
          "type": "string"
        },
        "accKy": {
          "format": "int32",
          "type": "integer"
        },
        "accTypes": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
    "AccLedgerReportRequestDto": {
      "type": "object",
      "properties": {
        "fromDate": {
          "format": "date",
          "type": "string"
        },
        "toDate": {
          "format": "date",
          "type": "string"
        },
        "accKy": {
          "format": "int32",
          "type": "integer"
        },
        "accTypKy": {
          "format": "int32",
          "type": "integer"
        },
        "multiCurrency": {
          "type": "boolean"
        },
        "landScope": {
          "type": "boolean"
        },
        "orderBy": {
          "type": "string"
        },
        "pageBreak": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "AccTrnDetailsReportRequestDto": {
      "type": "object",
      "properties": {
        "fromDate": {
          "format": "date",
          "type": "string"
        },
        "toDate": {
          "format": "date",
          "type": "string"
        },
        "fromTrnNo": {
          "format": "int32",
          "type": "integer"
        },
        "toTrnNo": {
          "format": "int32",
          "type": "integer"
        },
        "fromDocNo": {
          "type": "string"
        },
        "toDocNo": {
          "type": "string"
        },
        "fromYurRef": {
          "type": "string"
        },
        "toYurRef": {
          "type": "string"
        },
        "accKy": {
          "format": "int32",
          "type": "integer"
        },
        "accTypKy": {
          "format": "int32",
          "type": "integer"
        },
        "trnTypKy": {
          "format": "int32",
          "type": "integer"
        },
        "isContra": {
          "type": "boolean"
        },
        "isSummary": {
          "type": "boolean"
        },
        "isLandscape": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
    "ChqReturnReportRequestDto": {
      "type": "object",
      "properties": {
        "fromDate": {
          "format": "date",
          "type": "string"
        },
        "toDate": {
          "format": "date",
          "type": "string"
        },
        "accKy": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "GrossProfitItemLocationRequestDto": {
      "type": "object",
      "properties": {
        "fromDate": {
          "format": "date",
          "type": "string"
        },
        "toDate": {
          "format": "date",
          "type": "string"
        },
        "itmKy": {
          "format": "int32",
          "type": "integer"
        },
        "locKy": {
          "format": "int32",
          "type": "integer"
        },
        "buKy": {
          "format": "int32",
          "type": "integer"
        },
        "repAdrKy": {
          "format": "int32",
          "type": "integer"
        },
        "itmTypKy": {
          "format": "int32",
          "type": "integer"
        },
        "itmCat1Ky": {
          "format": "int32",
          "type": "integer"
        },
        "itmCat2Ky": {
          "format": "int32",
          "type": "integer"
        },
        "itmCat3Ky": {
          "format": "int32",
          "type": "integer"
        },
        "itmCat4Ky": {
          "format": "int32",
          "type": "integer"
        },
        "adrKy": {
          "format": "int32",
          "type": "integer"
        },
        "groupBy1": {
          "type": "string"
        },
        "groupBy2": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "ItemBatchReportRequestDto": {
      "type": "object",
      "properties": {
        "itmKy": {
          "format": "int32",
          "type": "integer"
        },
        "expiryDate": {
          "format": "date",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "ItmCat1WiseTrnSumRequestDto": {
      "type": "object",
      "properties": {
        "fromDate": {
          "format": "date",
          "type": "string"
        },
        "toDate": {
          "format": "date",
          "type": "string"
        },
        "locKy": {
          "format": "int32",
          "type": "integer"
        },
        "itmTypKy": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "NonPerformedCustomerRequestDto": {
      "type": "object",
      "properties": {
        "fromDate": {
          "format": "date",
          "type": "string"
        },
        "toDate": {
          "format": "date",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "NonPerformedItemRequestDto": {
      "type": "object",
      "properties": {
        "fromDate": {
          "format": "date",
          "type": "string"
        },
        "toDate": {
          "format": "date",
          "type": "string"
        },
        "itmCat1Ky": {
          "format": "int32",
          "type": "integer"
        },
        "itmCat2Ky": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "ReportSalesByRepRequestDto": {
      "type": "object",
      "properties": {
        "fromDate": {
          "format": "date-time",
          "type": "string"
        },
        "toDate": {
          "format": "date-time",
          "type": "string"
        },
        "repName": {
          "type": "string"
        },
        "itemKy": {
          "format": "int32",
          "type": "integer"
        },
        "salesReturn": {
          "type": "boolean"
        },
        "viewType": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "SalesDetailsByPaymentModeRequestDto": {
      "type": "object",
      "properties": {
        "fromDate": {
          "format": "date",
          "type": "string"
        },
        "toDate": {
          "format": "date",
          "type": "string"
        },
        "paymentModeName": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "SalesItemReportRequestDto": {
      "type": "object",
      "properties": {
        "fromDate": {
          "format": "date",
          "type": "string"
        },
        "toDate": {
          "format": "date",
          "type": "string"
        },
        "accKy": {
          "format": "int32",
          "type": "integer"
        },
        "itmKy": {
          "format": "int32",
          "type": "integer"
        },
        "salesOnly": {
          "type": "boolean"
        },
        "summary": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "SalesRepWiseCustomerRequestDto": {
      "type": "object",
      "properties": {
        "fromDate": {
          "format": "date",
          "type": "string"
        },
        "toDate": {
          "format": "date",
          "type": "string"
        },
        "repAdrKy": {
          "format": "int32",
          "type": "integer"
        },
        "itmKy": {
          "format": "int32",
          "type": "integer"
        },
        "onlySales": {
          "type": "boolean"
        },
        "isSummary": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
    "SetOffDetailsReportRequestDto": {
      "type": "object",
      "properties": {
        "fromDate": {
          "format": "date",
          "type": "string"
        },
        "toDate": {
          "format": "date",
          "type": "string"
        },
        "accKy": {
          "format": "int32",
          "type": "integer"
        },
        "varName": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "TrnDetReportRequestDto": {
      "type": "object",
      "properties": {
        "fromDate": {
          "format": "date",
          "type": "string"
        },
        "toDate": {
          "format": "date",
          "type": "string"
        },
        "trnTyp": {
          "type": "string"
        },
        "adrNm": {
          "type": "string"
        },
        "itmNm": {
          "type": "string"
        },
        "freeItemsOnly": {
          "type": "boolean"
        },
        "repAdrKy": {
          "format": "int32",
          "type": "integer"
        },
        "showDetails": {
          "type": "boolean"
        },
        "summary": {
          "type": "boolean"
        },
        "orderBy": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  }
}